Added accessor methods for the dataset metadata to AbstractDataset.
These define the metadata to be stored in the datacube's acquisition
and dataset tables.

Also added a get_geo_transform method which is needed to tile the data.

These accessors have docstrings but some are incomplete and they
need to be checked for accuracy by Alex at some stage.

modified:   abstract_dataset.py