Rename 'src' folder to match module name 'agdc'
The current repository layout prevents the codebase from being
run directly after checkout, as the package layout is inconsistent
with the module import path. A developer must either install the package
first or rename/link the source directory to run the code.

This also leads to an unusual testing environment: the tests import
agdc code using a different path to the code itself, leading
files to be imported twice under different prefixes, possibly
unexpectedly from different locations, which can lead
to unexpected results.

This commit renames the 'src' directory to 'agdc' to resolve these
issues. It also updates tests to import using an absolute 'agdc'
prefix, to be consistent with the code itself.

This repository layout is very common within GA and the wider Python
community.

Example online guides which recommend this layout:

    - http://www.kennethreitz.org/essays/repository-structure-and-python
    - http://www.scotttorborg.com/python-packaging/minimal.html
    - http://www.jeffknupp.com/blog/2013/08/16/open-sourcing-a-python-project-the-right-way/
    - https://stackoverflow.com/questions/193161/what-is-the-best-project-structure-for-a-python-application